requirements.txt
flask==2.2.5
flask-cors==4.0.0
flask-restful==0.3.10
# pydantic==2.5.0
pydantic>=2.3.0
# sqlalchemy>=1.4,<2.0
# sqlalchemy==1.4.46
# sqlalchemy==2.0.30
# sqlalchemy==1.3.24
sqlalchemy==1.4.46



# langchain==0.1.13
# langchain-community==0.0.36
# langchain-openai==0.2.1
# langchain==0.2.3
# langchain-community==0.0.38
# langchain-openai==0.2.2
# Compatible versions

# langchain==0.0.330
# langchain-community==0.0.21
# langchain-openai==0.0.8

# langchain==0.0.330
# langchain-community==0.0.18
# langchain-openai==0.0.92

#torch==2.3.0 --index-url https://download.pytorch.org/whl/cpu


psycopg2-binary==2.9.9
pymongo==4.6.0
# langchain==0.1.0
# langchain-openai==0.0.5

openai==1.10.0
google-cloud-texttospeech==2.15.0
kafka-python==2.0.2
apache-airflow==2.8.0
boto3==1.33.0
redis==5.0.1
celery==5.3.4
numpy==1.26.2
transformers==4.36.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
